requirements.txt
aiopg==1.4.0
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
exceptiongroup==1.1.1
fastapi==0.96.1
filelock==3.12.1
greenlet==2.0.2
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
load-dotenv==0.1.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
parsel==1.8.1
Protego==0.2.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
schedule==1.2.0
Scrapy==2.9.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
tldextract==3.4.4
Twisted==22.10.0
typing-extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
w3lib==2.1.1
zipp==3.15.0
zope.interface==6.0
